backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
alembic>=1.13.3
psycopg2-binary>=2.9.10
pydantic>=2.10.0
pydantic-settings>=2.6.0
bcrypt>=4.0.0
email-validator>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
apscheduler>=3.10.4
pandas>=2.2.3
openpyxl>=3.1.5
httpx>=0.27.0
